Client test cleanup
