+++ test/client_rebuild_test.go
@@ -1,0 +1,178 @@
+package test
+
+import (
+	"crypto/rand"
+	"fmt"
+	"os"
+	"testing"
+
+	"github.com/gadget-inc/dateilager/internal/auth"
+	util "github.com/gadget-inc/dateilager/internal/testutil"
+	"github.com/stretchr/testify/require"
+)
+
+func TestRebuild(t *testing.T) {
+	tc := util.NewTestCtx(t, auth.Project, 1)
+	defer tc.Close()
+
+	writeProject(tc, 1, 1)
+	writeObject(tc, 1, 1, nil, "a", "a v1")
+	writeObject(tc, 1, 1, nil, "b", "b v1")
+	writeObject(tc, 1, 1, nil, "c", "c v1")
+
+	c, _, close := createTestClient(tc)
+	defer close()
+
+	tmpDir := emptyTmpDir(t)
+	defer os.RemoveAll(tmpDir)
+
+	rebuild(tc, c, 1, nil, tmpDir, expectedResponse{
+		version: 1,
+		count:   3,
+	})
+
+	verifyDir(t, tmpDir, 1, map[string]expectedFile{
+		"a": {content: "a v1"},
+		"b": {content: "b v1"},
+		"c": {content: "c v1"},
+	})
+}
+
+func TestRebuildWithOverwritesAndDeletes(t *testing.T) {
+	tc := util.NewTestCtx(t, auth.Project, 1)
+	defer tc.Close()
+
+	writeProject(tc, 1, 2)
+	writeObject(tc, 1, 1, i(2), "a", "a v1 - long buffer of content")
+	writeObject(tc, 1, 1, i(2), "b", "b v1")
+	writeObject(tc, 1, 1, nil, "c", "c v1")
+	writeObject(tc, 1, 1, i(2), "e", "e v1")
+	writeObject(tc, 1, 2, nil, "a", "a v2")
+	writeObject(tc, 1, 2, nil, "d", "d v2")
+	writeSymlink(tc, 1, 2, nil, "e", "a")
+
+	c, _, close := createTestClient(tc)
+	defer close()
+
+	tmpDir := writeTmpFiles(t, 1, map[string]string{
+		"a": "a v1 - long buffer of content",
+		"b": "b v1",
+		"c": "c v1",
+		"e": "e v1",
+	})
+	defer os.RemoveAll(tmpDir)
+
+	rebuild(tc, c, 1, nil, tmpDir, expectedResponse{
+		version: 2,
+		count:   4,
+	})
+
+	verifyDir(t, tmpDir, 2, map[string]expectedFile{
+		"a": {content: "a v2"},
+		"c": {content: "c v1"},
+		"d": {content: "d v2"},
+		"e": {content: "a v2"},
+	})
+}
+
+func TestRebuildWithEmptyDirAndSymlink(t *testing.T) {
+	tc := util.NewTestCtx(t, auth.Project, 1)
+	defer tc.Close()
+
+	writeProject(tc, 1, 2)
+	writeObject(tc, 1, 1, nil, "a", "a v1")
+	writeObject(tc, 1, 1, nil, "d/e", "e v1")
+	writeEmptyDir(tc, 1, 1, nil, "b/")
+	writeSymlink(tc, 1, 2, nil, "c", "a")
+	writeSymlink(tc, 1, 2, nil, "f/g/h", "d/e")
+
+	c, _, close := createTestClient(tc)
+	defer close()
+
+	tmpDir := emptyTmpDir(t)
+	defer os.RemoveAll(tmpDir)
+
+	rebuild(tc, c, 1, nil, tmpDir, expectedResponse{
+		version: 2,
+		count:   5,
+	})
+
+	verifyDir(t, tmpDir, 2, map[string]expectedFile{
+		"a":     {content: "a v1"},
+		"d/e":   {content: "e v1"},
+		"b/":    {content: "", fileType: typeDirectory},
+		"c":     {content: "a", fileType: typeSymlink},
+		"f/g/h": {content: "d/e", fileType: typeSymlink},
+	})
+}
+
+func TestRebuildWithUpdatedEmptyDirectories(t *testing.T) {
+	tc := util.NewTestCtx(t, auth.Project, 1)
+	defer tc.Close()
+
+	writeProject(tc, 1, 1)
+	writeEmptyDir(tc, 1, 1, nil, "a/")
+	writeEmptyDir(tc, 1, 1, nil, "b/")
+
+	c, fs, close := createTestClient(tc)
+	defer close()
+
+	tmpDir := emptyTmpDir(t)
+	defer os.RemoveAll(tmpDir)
+
+	rebuild(tc, c, 1, nil, tmpDir, expectedResponse{
+		version: 1,
+		count:   2,
+	})
+
+	verifyDir(t, tmpDir, 1, map[string]expectedFile{
+		"a/": {content: "", fileType: typeDirectory},
+		"b/": {content: "", fileType: typeDirectory},
+	})
+
+	updateStream := newMockUpdateServer(tc.Context(), 1, map[string]expectedObject{
+		"a/c": {content: "a/c v2"},
+	})
+
+	err := fs.Update(updateStream)
+	require.NoError(t, err, "fs.Update")
+
+	rebuild(tc, c, 1, nil, tmpDir, expectedResponse{
+		version: 2,
+		count:   1,
+	})
+
+	verifyDir(t, tmpDir, 2, map[string]expectedFile{
+		"a/c": {content: "a/c v2"},
+		"b/":  {content: "", fileType: typeDirectory},
+	})
+}
+
+func TestRebuildWithManyObjects(t *testing.T) {
+	tc := util.NewTestCtx(t, auth.Project, 1)
+	defer tc.Close()
+
+	expectedFiles := make(map[string]expectedFile)
+
+	writeProject(tc, 1, 1)
+	for i := 0; i < 500; i++ {
+		bytes := make([]byte, 50000)
+		_, err := rand.Read(bytes)
+		require.NoError(t, err, "could not generate random bytes")
+		writeObject(tc, 1, 1, nil, fmt.Sprintf("/%d", i), string(bytes))
+		expectedFiles[fmt.Sprintf("/%d", i)] = expectedFile{content: string(bytes)}
+	}
+
+	c, _, close := createTestClient(tc)
+	defer close()
+
+	tmpDir := emptyTmpDir(t)
+	defer os.RemoveAll(tmpDir)
+
+	rebuild(tc, c, 1, nil, tmpDir, expectedResponse{
+		version: 1,
+		count:   500,
+	})
+
+	verifyDir(t, tmpDir, 1, expectedFiles)
+}
